+++ src/components/CardInfoForm/CardExpiredDate/index.tsx
@@ -4,12 +4,44 @@ import useInputValidation from "../../../hooks/useInputValidation";
 import { ExpiredDate } from "../../../types";
 import Input from "../../../common/Input";
 import InputContainer from "../../../common/InputContainer";
-import { checkExpiredDate } from "../../../validations/cardInfoForm";
 import { Context } from "../../../contexts/CardContext";
 
 interface CardExpiredDateProps {
-  validateFormValidation: any;
+  validateFormValidation: (key: string, isValid: boolean) => void;
 }
+
+const checkExpiredDate = (expiredDate: ExpiredDate) => {
+  const date = new Date();
+  const currentYear = date.getFullYear() % 100;
+  const currentMonth = date.getMonth() + 1;
+  const targetYear = Number(expiredDate.year);
+  const targetMonth = Number(expiredDate.month);
+
+  if (!Object.keys(expiredDate).every(key => expiredDate[key].length === 2)) {
+    throw new Error("MM/YY 형태로 입력해주세요. (예: 01/23)");
+  }
+
+  if (targetMonth < 1 || targetMonth > 12) {
+    throw new Error("1월부터 12월 사이의 값을 입력해주세요.");
+  }
+
+  if (targetYear < currentYear || (targetYear === currentYear && targetMonth < currentMonth))
+    throw new Error(
+      `해당 만료일은 이미 지났습니다. (현재: ${currentMonth
+        .toString()
+        .padStart(2, "0")}/${currentYear})`
+    );
+
+  if (
+    targetYear > currentYear + 5 ||
+    (targetYear === currentYear + 5 && targetMonth > currentMonth)
+  )
+    throw new Error(
+      `만료일은 현재로부터 5년이내여야합니다.  (현재: ${currentMonth
+        .toString()
+        .padStart(2, "0")}/${currentYear})`
+    );
+};
 
 export default function CardExpiredDate({ validateFormValidation }: CardExpiredDateProps) {
   const { inputValidation, validateInput, isValidInput } = useInputValidation(false);
